requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
espn-api==0.22.1
et-xmlfile==1.1.0
Flask==2.0.2
fonttools==4.37.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.4.4
MarkupSafe==2.0.1
matplotlib==3.5.3
mdutils==1.4.0
numpy==1.21.3
openai==0.19.0
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pandas-stubs==1.2.0.35
Pillow==9.2.0
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
seaborn==0.12.0
six==1.16.0
tabulate==0.8.10
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
